+++ universe/ConditionSource.h
@@ -18,7 +18,7 @@ struct FO_COMMON_API Source final : public Condition {
     void GetDefaultInitialCandidateObjects(const ScriptingContext& parent_context,
                                            ObjectSet& condition_non_targets) const override;
     [[nodiscard]] std::string Description(bool negated = false) const override;
-    [[nodiscard]] std::string Dump(unsigned short ntabs = 0) const override;
+    [[nodiscard]] std::string Dump(uint8_t ntabs = 0) const override;
     void SetTopLevelContent(const std::string& content_name) override {}
     [[nodiscard]] unsigned int GetCheckSum() const override;
