+++ universe/Fighter.cpp
@@ -43,7 +43,7 @@ const std::string& Fighter::SpeciesName() const
 void Fighter::SetDestroyed(bool destroyed)
 { m_destroyed = destroyed; }
 
-std::string Fighter::Dump(unsigned short ntabs) const {
+std::string Fighter::Dump(uint8_t ntabs) const {
     std::stringstream os;
     os << UniverseObject::Dump(ntabs);
     os << " (Combat Fighter) damage: " << m_damage;
