+++ universe/Special.h
@@ -46,7 +46,7 @@ public:
 
     auto&       Name() const            { return m_name; }          ///< returns the unique name for this type of special
     std::string Description() const;                                ///< returns a text description of this type of special
-    std::string Dump(unsigned short ntabs = 0) const;               ///< returns a data file format representation of this object
+    std::string Dump(uint8_t ntabs = 0) const;               ///< returns a data file format representation of this object
     auto*       Stealth() const         { return m_stealth.get(); } ///< returns the stealth of the special, which determines how easily it is seen by empires
     auto&       Effects() const         { return m_effects; }               ///< returns the EffectsGroups that encapsulate the effects that specials of this type have.
     float       SpawnRate() const       { return m_spawn_rate; }
